+++ src/renderer/ipc/settingsIpcHandler.ts
@@ -2,24 +2,6 @@ import { ipcRenderer } from 'electron';
 import { AppSetting } from '../../main/generated/client';
 import { AppSettingUpdatePropsInterface } from '../../main/controllers/SettingsController';
 
-/**
- * @deprecated in favor of updateAppSettings
- */
-export async function updateFrontEndTheme(theme: string): Promise<void> {
-  const updated = await ipcRenderer.invoke('updateFrontEndTheme', theme);
-
-  return updated;
-}
-
-/**
- * @deprecated in favor of getSetting
- */
-export async function getFrontEndTheme(): Promise<string> {
-  const theme = await ipcRenderer.invoke('getFrontEndTheme');
-
-  return theme.value;
-}
-
 export async function updateAppSettings(
   settings: AppSettingUpdatePropsInterface[]
 ): Promise<AppSetting[]> {