+++ src/main/utils/logRotation.ts
@@ -2,6 +2,7 @@
 import path from 'path';
 import * as fs from 'fs';
 import log from 'electron-log';
+import compressing from 'compressing';
 import { isDevelopment } from './globalConstants';
 import getAppDataFolder from './fsUtils';
 
@@ -9,7 +10,7 @@ import getAppDataFolder from './fsUtils';
  * Archives the app log file on a daily basis with a custom name, preventing it from being too large.
  * @since 0.1.0
  */
-export default function rotateLogFile(): void {
+export default async function rotateLogFile(): Promise<void> {
   try {
     const today = new Date();
     const filePath = path.resolve(
@@ -45,6 +46,13 @@ export default function rotateLogFile(): void {
       fs.writeFileSync(archiveFilePath, logContent);
 
       fs.writeFileSync(filePath, '');
+
+      await compressing.zip.compressFile(
+        archiveFilePath,
+        archiveFilePath.replace('.log', '.zip')
+      );
+
+      fs.rmSync(archiveFilePath);
 
       log.info(
         `Previous log file has been archived as ${archiveFilePath} due to file rotation`