+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added a configurable minimize to system tray feature
 - Deprecated the updateFrontEndTheme method in favor of the generic updateSettings method
+- Fixed log rotation file name
 
 # 0.3.0